update create post screen structure
