+++ modules/screen-create-post/assets/camera.png
